Adds the marking of duplicate games. Fixes #6
Lets the user mark a game as a duplicate. If the game is a duplicate,
then it doesn't show up in search results. I can then look to see if the
duplicate game is valid or not, and if so, I can delete it (hopefully I
can automate this process sometime in the future).

Given that I need to tell which of a (possible) series of games is the
'correct' choice (which I will most likely define as the one people
actually upload pictures to), I also added the ability to view pictures
from the games page in the admin interface. This will let me see at a
glance which of the multiple copies of the game should be the one to
show up in search results.