Adds consoles to admin panel.
Fixes #12 as well. In the process of doing this, I realized the issue
with unchecking duplicates in the admin panel. The problem was that
duplicate was not set as attr_accessible, so when activeadmin tried to
mass assign it it would fail. This commit fixes that issue as well
(since it needed to make console_id accessible also).