Fix issue with new version of will_paginate
The new version of will_paginate removed support for paginating arrays, it required that you have an collection that explicitly supports pagination. Sunspot collections actually fit that bill, but since I was sorting the results manually (why would I ever do that???) it was turning it into an array, removing that functionality.

So this commit is basically to remove my custom sorting of search results. Again, I don't know why I did that in the first place, as the whole point of the search is to give me relevant results in order. I don't know why I thought I knew better than Solr...