Fix ConsoleGrid 500'ing when given a console that it didnt know.
For legacy reasons, ConsoleGrid assumed that the "console" parameter in the api call would always exist. This is clearly not the case, and it would 500 when given an odd console.

Instead of 500'ing, now it returns a 204 no content, which is what we do when there are no pictures for a game. Now Ice wont show a big scary error message when the user does something silly.